requirements.txt
fastapi
uvicorn
python-multipart
pydantic
PyMuPDF
python-docx
spacy
sentence-transformers
numpy
pandas